The scripts and makefile have been added to git.  The source files are now in ./src and the include files are in the include directory.
The main script is running the software cfd and is working well.
It now has the compatability with process script and requires a second number in addition to run number when calling for many files

./EvtBuilder runNumber [numberOfInputFiles]

If number of input files is 1 it can be ommited

Need to change the metta script account for many input files.