Changed the correlated event triggering conditions.  Instead of the complicated searches in a window of 6
it is now restricted to a window of 4 and sorts the information based on channel number.  This aviods the
necessity to check to make sure you are subtracting the same PMT each time.  The main conditions are:
All 4 channels have to be present in the set of 4.  The average of the two bar's times has to be less then
10 clock tics