The corections run mode is now in place.  It will take the previous output files as an inputfile
for a second building with corrections.  The first time you run the builder makes the normal tree but now
has the correlatedtimes branch we keeps the times of the groups of four identified on the first run.  also
keeps there energies.  Note only internal timing will work in this mode as the tree no longer has the traces
in it.  No input checking for this right now.  Need to do this at one point.

Also maybe i should have it only look for a corrections file when it is in a corrections run mode

Had to add a variety of methods to all the manager classes to get this feature to work